v0.1.0
Basic code for reading from database.txt file.

The program sets the current directory to the folder that both main.py and database.txt are in, so it can use absolute paths to refer to the .txt file. This way it should be able to work on any computer. The program also calls the lines from the database in a random order every time, so the words don't always come in the same order.